backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymongo==4.6.1
python-dotenv==1.0.1
pydantic==2.6.4
